added logger and messages as well as some validation

+++ src/main/java/com/acn/gamechangers/mymusicranking/controller/SongController.java
@@ -3,20 +3,24 @@ package com.acn.gamechangers.mymusicranking.controller;
 import com.acn.gamechangers.mymusicranking.model.Album;
 import com.acn.gamechangers.mymusicranking.model.Category;
 import com.acn.gamechangers.mymusicranking.model.Song;
-import com.acn.gamechangers.mymusicranking.repository.AlbumRepository;
-import com.acn.gamechangers.mymusicranking.repository.CategoryRepository;
-import com.acn.gamechangers.mymusicranking.repository.SongRepository;
+import com.acn.gamechangers.mymusicranking.services.AlbumServiceImpl;
+import com.acn.gamechangers.mymusicranking.services.CategoryServiceImpl;
 import com.acn.gamechangers.mymusicranking.services.SongServiceImpl;
 import lombok.extern.log4j.Log4j2;
+import org.apache.logging.log4j.Level;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
+
+import static com.acn.gamechangers.mymusicranking.config.Helper.getPreviousPageByRequest;
 
 @Controller()
 @RequestMapping("/songs")
@@ -24,15 +28,13 @@ import java.util.List;
 public class SongController {
 
     private final SongServiceImpl songService;
-    private final CategoryRepository categoryRepository;
-    private final AlbumRepository albumRepository;
-    private final SongRepository songRepository;
+    private final CategoryServiceImpl categoryService;
+    private final AlbumServiceImpl albumService;
 
-    public SongController(SongServiceImpl songService, CategoryRepository categoryRepository, AlbumRepository albumRepository, SongRepository songRepository) {
+    public SongController(SongServiceImpl songService, CategoryServiceImpl categoryService, AlbumServiceImpl albumService) {
         this.songService = songService;
-        this.categoryRepository = categoryRepository;
-        this.albumRepository = albumRepository;
-        this.songRepository = songRepository;
+        this.categoryService = categoryService;
+        this.albumService = albumService;
     }
 
     @GetMapping("/form")
@@ -40,19 +42,20 @@ public class SongController {
         ModelAndView modelAndView = new ModelAndView();
         Song song = new Song();
         song.setReleaseDate(LocalDate.now());
-        List<Category> categoryList = categoryRepository.findAll();
-        List<Album> albumList = albumRepository.findAll();
+        List<Category> categoryList = categoryService.getCategoryList();
+        List<Album> albumList = albumService.getAlbumList();
         modelAndView.addObject("categoryList", categoryList);
         modelAndView.addObject("albumList", albumList);
         modelAndView.addObject("song", song);
         modelAndView.setViewName("form");
+        log.log(Level.INFO, "");
         return modelAndView;
     }
 
     @GetMapping("")
     public ModelAndView readSongs() {
         ModelAndView modelAndView = new ModelAndView();
-        List<Song> songList = songRepository.findAll();
+        List<Song> songList = songService.getSongList();
         modelAndView.addObject("songList", songList);
         modelAndView.setViewName("songs");
         return modelAndView;
@@ -62,29 +65,50 @@ public class SongController {
     public ModelAndView addSong(@Valid Song song, BindingResult bindingResult) {
         ModelAndView modelAndView = new ModelAndView();
         modelAndView.addObject("song", song);
-        List<Category> categoryList = categoryRepository.findAll();
-        List<Album> albumList = albumRepository.findAll();
+        List<Category> categoryList = categoryService.getCategoryList();
+        List<Album> albumList = albumService.getAlbumList();
         modelAndView.addObject("categoryList", categoryList);
         modelAndView.addObject("albumList", albumList);
-        modelAndView.addObject("bindingResult", bindingResult);
         modelAndView.setViewName("form");
         if (bindingResult.hasErrors()) {
+            modelAndView.addObject("message", "Please correct the errors in form!");
+            log.info("There where errors in the form!");
+            modelAndView.addObject("bindingResult", bindingResult);
             return modelAndView;
+        } else {
+           /* List<Album> songAlbums = song.getAlbumList();
+            List<Category> songCategories = song.getCategoryList();
+            for (Album album : songAlbums) {
+                album.addSongAlbum(song);
+                albumService.addAlbum(album);
+            }
+            for (Category category : songCategories) {
+                category.addSongCategory(song);
+                categoryService.addCategory(category);
+            }*/
+            
+            songService.addSong(song);
+            modelAndView.addObject("message",
+                    "Song got registered successfully! Song Title: " + song.getTitle());
         }
-        songService.addSong(song);
-        modelAndView.addObject("message",
-                "Song got registered successfully! Song Title: " + song.getTitle());
         return modelAndView;
     }
 
     @PostMapping("/deleteSong")
-    public ModelAndView removeSong(@RequestParam Long songId) {
-
-        songService.remove(songId);
+    public ModelAndView removeSong(@RequestParam Long songId, HttpServletRequest request) {
         ModelAndView modelAndView = new ModelAndView();
-        List<Song> songList = songRepository.findAll();
-        modelAndView.addObject("songList", songList);
-        modelAndView.setViewName("songs");
+        modelAndView.setViewName(getPreviousPageByRequest(request).orElse("/songs"));
+        Optional<Song> byIdOptional = songService.findByIdOptional(songId);
+        if (byIdOptional.isEmpty()) {
+            modelAndView.addObject("message", "There is no Song with this songId: " + songId);
+            log.info("There is no Song with this songId: " + songId);
+        } else {
+            songService.remove(songId);
+            modelAndView.addObject("message", "Song with songId" + songId + " got removed!");
+            log.info("Song with songId" + songId + " got removed!");
+            List<Song> songList = songService.getSongList();
+            modelAndView.addObject("songList", songList);
+        }
         return modelAndView;
     }
